refactor out replace fixture
