+++ shared-data/js/helpers/deckConfiguration/getFixtureReplacment.ts
@@ -1,0 +1,224 @@
+import { isEqual } from 'lodash'
+
+import {
+  ABSORBANCE_READER_V1_FIXTURE,
+  DEFAULT_AA_FOR_WASTE_CHUTE,
+  FAKE_STAGING_AREA_RIGHT_SLOT,
+  FAKE_STAGING_SLOT_WITH_MAG_BLOCK_FIXTURE,
+  FAKE_WASTE_CHUTE_WITH_EMPTY_SLOT_FIXTURE,
+  FLEX_ROBOT_TYPE,
+  FLEX_STACKER_V1_FIXTURE,
+  FLEX_STACKER_WITH_MAG_BLOCK_FIXTURE,
+  MAGNETIC_BLOCK_V1,
+  SINGLE_CENTER_SLOT_FIXTURE,
+  SINGLE_LEFT_CUTOUTS,
+  SINGLE_LEFT_SLOT_FIXTURE,
+  SINGLE_RIGHT_CUTOUTS,
+  SINGLE_RIGHT_SLOT_FIXTURE,
+  STAGING_AREA_RIGHT_SLOT_FIXTURE,
+  WASTE_CHUTE_FIXTURES,
+  WASTE_CHUTE_FLEX_STACKER_FIXTURES,
+  WASTE_CHUTE_RIGHT_ADAPTER_NO_COVER_FIXTURE,
+  WASTE_CHUTE_WITH_FAKE_FIXTURES,
+} from '../../constants'
+import {
+  getAAByAAId,
+  getAAsToFixtureIdFromDeckDefWithFakes,
+  getAAWithFakesFromCutoutFixtureId,
+  getAAWithFakesFromVSId,
+  getVisualSlotIdForAA,
+} from '../../fixtures'
+import { getDeckDefFromRobotType } from '../../helpers'
+
+import type { CutoutFixtureId, CutoutId } from '../../../deck'
+import type {
+  AddressableAreaNamesWithFakes,
+  CutoutFixtureIdsWithFakes,
+} from '../../constants'
+import type { CutoutConfig, CutoutConfigMap, DeckDefinition } from '../../types'
+
+export const getCutoutFixtureReplacementIfNeeded = (
+  cutoutFixtureId: CutoutFixtureId,
+  cutoutId: CutoutId,
+  deckDefinition: DeckDefinition
+): CutoutFixtureIdsWithFakes => {
+  if (SINGLE_RIGHT_CUTOUTS.includes(cutoutId)) {
+    if (
+      cutoutFixtureId === SINGLE_RIGHT_SLOT_FIXTURE &&
+      deckDefinition.robot.model === FLEX_ROBOT_TYPE
+    ) {
+      return FAKE_STAGING_AREA_RIGHT_SLOT
+    } else if (
+      cutoutFixtureId === WASTE_CHUTE_RIGHT_ADAPTER_NO_COVER_FIXTURE &&
+      deckDefinition.robot.model === FLEX_ROBOT_TYPE
+    ) {
+      return FAKE_WASTE_CHUTE_WITH_EMPTY_SLOT_FIXTURE
+    } else if (cutoutFixtureId === MAGNETIC_BLOCK_V1) {
+      return 'fakeStagingSlotWithMagBlockV1'
+    }
+  }
+  return cutoutFixtureId
+}
+
+export const getReplacementFixtureForFixtureRemoval = (
+  cutoutFixtureId: CutoutFixtureIdsWithFakes,
+  cutoutId: CutoutId,
+  addressableAreaId?: AddressableAreaNamesWithFakes // nullable for PD
+): CutoutFixtureId => {
+  if (cutoutFixtureId === STAGING_AREA_RIGHT_SLOT_FIXTURE) {
+    return SINGLE_RIGHT_SLOT_FIXTURE
+  } else if (addressableAreaId && SINGLE_RIGHT_CUTOUTS.includes(cutoutId)) {
+    const cutoutFixtureReplacment = replaceCutoutFixtureForFixtureRemoval(
+      cutoutFixtureId,
+      cutoutId,
+      addressableAreaId
+    )
+    return getReplacementFixtureForFakeFixture(cutoutFixtureReplacment)
+  } else if (SINGLE_RIGHT_CUTOUTS.includes(cutoutId)) {
+    return SINGLE_RIGHT_SLOT_FIXTURE
+  } else if (SINGLE_LEFT_CUTOUTS.includes(cutoutId)) {
+    return SINGLE_LEFT_SLOT_FIXTURE
+  }
+  return SINGLE_CENTER_SLOT_FIXTURE
+}
+
+/**
+ * Given a fake cutout fixutre find a replacment to store on the server
+ * @param cutoutFixtureId: a cutoutId we wish to replace with a non fake fixture
+ * @returns the relevant cutoutFixtureId to store on the server
+ */
+export const getReplacementFixtureForFakeFixture = (
+  cutoutFixtureId: CutoutFixtureIdsWithFakes
+): CutoutFixtureId => {
+  if (cutoutFixtureId === FAKE_STAGING_AREA_RIGHT_SLOT) {
+    return SINGLE_RIGHT_SLOT_FIXTURE
+  }
+  if (cutoutFixtureId === FAKE_WASTE_CHUTE_WITH_EMPTY_SLOT_FIXTURE) {
+    return WASTE_CHUTE_RIGHT_ADAPTER_NO_COVER_FIXTURE
+  }
+  if (cutoutFixtureId === FAKE_STAGING_SLOT_WITH_MAG_BLOCK_FIXTURE) {
+    return MAGNETIC_BLOCK_V1
+  }
+  return cutoutFixtureId
+}
+/**
+ * Given a list of cutout configs replace fixtures with their fake fixtures
+ * @param: a list of cutout config from the server
+ * @returns: a list of cutout config with relevant AA
+ */
+export const replaceFixtureToFakeFixtureAndTransformCutoutFixturesToAA = (
+  cutoutFixtures: CutoutConfig[]
+): CutoutConfigMap[] => {
+  const deckDef = getDeckDefFromRobotType('OT-3 Standard')
+  return cutoutFixtures.reduce<CutoutConfigMap[]>((acc, obj) => {
+    const cutoutFixtureReplacment = getCutoutFixtureReplacementIfNeeded(
+      obj.cutoutFixtureId,
+      obj.cutoutId,
+      deckDef
+    )
+    const aaPerCutoutFixture = getAAWithFakesFromCutoutFixtureId(
+      obj.cutoutId,
+      cutoutFixtureReplacment ?? obj.cutoutFixtureId,
+      deckDef
+    )
+
+    if (WASTE_CHUTE_FIXTURES.includes(obj.cutoutFixtureId)) {
+      acc.push({
+        ...obj,
+        addressableAreaId: DEFAULT_AA_FOR_WASTE_CHUTE,
+        cutoutFixtureId: cutoutFixtureReplacment,
+      })
+      const otherWasteChuteAA = aaPerCutoutFixture?.find(
+        aa => getAAByAAId(aa, deckDef).areaType !== 'wasteChute'
+      )
+      if (otherWasteChuteAA != null) {
+        acc.push({
+          ...obj,
+          addressableAreaId: otherWasteChuteAA,
+          cutoutFixtureId: cutoutFixtureReplacment,
+        })
+      }
+    } else if (obj.cutoutFixtureId === ABSORBANCE_READER_V1_FIXTURE) {
+      const absorbanceReaderAA = aaPerCutoutFixture?.find(
+        aa => getAAByAAId(aa, deckDef).areaType === 'absorbanceReader'
+      )
+      if (absorbanceReaderAA != null) {
+        acc.push({
+          ...obj,
+          addressableAreaId: absorbanceReaderAA,
+          cutoutFixtureId: cutoutFixtureReplacment,
+        })
+      }
+    } else {
+      aaPerCutoutFixture?.forEach(item => {
+        acc.push({
+          ...obj,
+          addressableAreaId: item,
+          cutoutFixtureId: cutoutFixtureReplacment,
+        })
+      })
+    }
+    return acc
+  }, [])
+}
+
+/**
+ * Given a cutout fixture to remove find what fixture to use instead
+ * @param cutoutFixtureRemoved: fixtures to remove
+ * @param cutoutId
+ * @param addressableAreaId to remove from fixture
+ * @returns fixture to replace with
+ */
+export const replaceCutoutFixtureForFixtureRemoval = (
+  cutoutFixtureRemoved: CutoutFixtureIdsWithFakes,
+  cutoutId: CutoutId,
+  addressableAreaId: AddressableAreaNamesWithFakes
+): CutoutFixtureIdsWithFakes => {
+  const deckDef = getDeckDefFromRobotType('OT-3 Standard')
+  const aaPerFixtureOptions = getAAsToFixtureIdFromDeckDefWithFakes(
+    cutoutId,
+    deckDef
+  )
+  const aaForFixtureRemoval = getAAWithFakesFromCutoutFixtureId(
+    cutoutId,
+    cutoutFixtureRemoved,
+    deckDef
+  )
+  if (WASTE_CHUTE_WITH_FAKE_FIXTURES.includes(cutoutFixtureRemoved)) {
+    if (addressableAreaId === DEFAULT_AA_FOR_WASTE_CHUTE) {
+      return WASTE_CHUTE_FLEX_STACKER_FIXTURES.includes(cutoutFixtureRemoved)
+        ? FLEX_STACKER_V1_FIXTURE
+        : SINGLE_RIGHT_SLOT_FIXTURE
+    } else {
+      return WASTE_CHUTE_RIGHT_ADAPTER_NO_COVER_FIXTURE
+    }
+  } else if (cutoutFixtureRemoved === ABSORBANCE_READER_V1_FIXTURE) {
+    return SINGLE_RIGHT_SLOT_FIXTURE
+  } else if (cutoutFixtureRemoved === FLEX_STACKER_V1_FIXTURE) {
+    return SINGLE_RIGHT_SLOT_FIXTURE
+  } else if (cutoutFixtureRemoved === FLEX_STACKER_WITH_MAG_BLOCK_FIXTURE) {
+    if (getAAByAAId(addressableAreaId, deckDef).areaType === 'flexStacker') {
+      return FAKE_STAGING_SLOT_WITH_MAG_BLOCK_FIXTURE
+    } else {
+      return FLEX_STACKER_V1_FIXTURE
+    }
+  } else {
+    const updated = aaForFixtureRemoval?.map(aa => {
+      const vsId = getVisualSlotIdForAA(cutoutId, cutoutFixtureRemoved, aa)
+      return aa === addressableAreaId ? getAAWithFakesFromVSId(vsId) : aa
+    })
+    const match = Object.entries(aaPerFixtureOptions).find(([, value]) => {
+      return isEqual(
+        value.sort(),
+        WASTE_CHUTE_WITH_FAKE_FIXTURES.includes(cutoutFixtureRemoved)
+          ? aaForFixtureRemoval?.sort()
+          : updated?.sort()
+      )
+    })
+    if (match) {
+      return match[0] as CutoutFixtureIdsWithFakes
+    }
+    // Fallback if no match found
+    return cutoutFixtureRemoved
+  }
+}
